refactor: wandb subpackage, finalize instead of final summary
Split the W&B monitor into monitors/wandb/{base,overview} for
readability. Rename save_final_summary to finalize - its real job is
finalizing the platform run (with the last metrics as summary); the
local final_summary.json files had no consumers and are no longer
written. Drop close() from the interface and registry: nothing called
it, and PrimeMonitor's shutdown (mark unfinalized runs failed, stop the
background loop) lives in __del__.
